Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -1,9 +1,9 @@
 #include <stdio.h>
 int main() {
-    float r,pie=3.14,area;
-    scanf("%f",r);
+    double r,pie=3.14,area;
+    scanf("%lf",r);
     area=pie*r*r;
-    printf("Area: %.2f\n",area);
+    printf("Area: %.2lf\n",area);
     
     
     return 0;
